refactors, waiting for kv interface
