+++ package.json
@@ -87,7 +87,7 @@
     "redis": "^4.6.4",
     "safe-stable-stringify": "^2.4.3",
     "snarkjs": "^0.6.1",
-    "warp-contracts": "^1.4.1",
+    "warp-contracts": "^1.4.2",
     "warp-contracts-lmdb": "^1.1.9",
     "warp-contracts-plugin-deploy": "^1.0.4",
     "warp-contracts-plugin-ethers": "^1.0.7",
