+++ tests/hollowdb.test.ts
@@ -4,7 +4,9 @@ import {DeployPlugin} from 'warp-contracts-plugin-deploy';
 import initialState from '../common/initialState';
 import fs from 'fs';
 import path from 'path';
-import {SDK, Admin, Prover, computeKey} from '../src';
+import {SDK, Admin} from '../src';
+import {Prover} from './utils/prover';
+import {computeKey} from './utils/computeKey';
 import type {CacheType} from '../src/sdk/types';
 import {randomBytes} from 'crypto';
 import constants from './constants';
